fixed javadoc
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@3607 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zuist-engine/trunk/src/main/java/fr/inria/zuist/engine/SceneManager.java
@@ -948,7 +948,7 @@ public class SceneManager implements CameraListener {
     
     /** Enable/disable level updating.
      * Calls to updateLevel(altitude) have no effect if level updating is disabled.
-     *@see #updateLevel(int layerIndex, double[] cameraBounds, float altitude)
+     *@see #updateLevel(int layerIndex, double[] cameraBounds, double altitude)
      */
     public void setUpdateLevel(boolean b){
 	updateLevel = b;
